FlowControl과 PLC 간 값 읽고 쓰기 기능 추가 (인덱스 확인 필요)

+++ SapphireXR/SapphireXR_App/ViewModels/FlowControllerViewModel.cs
@@ -28,7 +28,7 @@ namespace SapphireXR_App.ViewModels
                 }
                 if (controlValues.rampTime != null)
                 {
-                    PLCService.WriteRampTime(ControllerID, (int)controlValues.rampTime);
+                    PLCService.WriteRampTime(ControllerID, (Int16)controlValues.rampTime);
                 }
             });
         }
